Move to shared tests functions.
